refactor: deprecate Grid and GridArea
