+++ src/components/SharedComponents/Grid.deprecated/index.ts
@@ -1,0 +1,4 @@
+export { default } from "./Grid.deprecated";
+export type { GridProps } from "./Grid.deprecated";
+export type { GridAreaProps } from "./GridArea.deprecated";
+export { default as GridArea } from "./GridArea.deprecated";
