--- src/components/SharedComponents/Grid/index.ts
@@ -1,4 +1,0 @@
-export { default } from "./Grid";
-export type { GridProps } from "./Grid";
-export type { GridAreaProps } from "./GridArea";
-export { default as GridArea } from "./GridArea";
